Move testing gamelogic to new class
