+++ src/main.py
@@ -1,8 +1,8 @@
 import pygame
 from engine import engine, window
-from gameLogic import GameLogic
+from game import Game
 
 window = window.Window((800,600), "My game")
-gameLogic = GameLogic(window)
-engine = engine.Engine(gameLogic, window, max_fps=30)
+game = Game(window)
+engine = engine.Engine(game, window, max_fps=30)
 engine.run()
